New import structure for NLPDifferentiatior
